Update 26.1.md
The definition of flow f' in G' should be exactly the same as f in G;
except for two edges (u, x), and (x, v) which is equal to (u, v). On the
other hand, f(u, x) and f(x, v) are undefined, because G has no vertex x
which is augmented to G to construct G'.