Update 2.1.md
Improve 2.1-3 explanation

Improve the loop invariant proof with more precise explanations as shown in the book.
It's important to conclude that the algorithm is correct as inferred from the termination step, as that is one of the goals of using loop invariant.
Also note that the loop invariant does not need to hold true at every step of the loop iteration, only before and after of each loop iteration.