Update 16.1.md
O(n . log n) algorithm for the exercise 16.1-5 is described and proved correct. Original solution was in cubic time. Please write me if you have anything to fix in my proposal. I can make it even more formal if-need-to-be. I tried to keep it not too much formal, readable also for a non-mathematician.