Update 22.2.md
The two images order don't correspond to the text. 

The text says "Let G be the graph shown in the first picture, G_pi be the graph shown in the second picture". But the first picture is https://github.com/walkccc/CLRS/blob/master/docs/img/22.2-6-1.png which is G_pi.

Switching the order of the images